requirements.txt
PyQt6
rembg[gpu]
Pillow